refactor(tests): reorganize SQL integration tests and update configurations
- Moved SQL integration test files to a new directory structure under `helpers`.
- Renamed test files to use the `.sqlserver.test.ts` suffix for clarity.
- Removed outdated test files related to requirements specifications and pagination.
- Introduced a new global setup script for resetting the SQL Server database and running migrations.
- Updated the Vitest configuration to reflect the new file structure and naming conventions.
- Added new tests for norm reference mutations and requirements lifecycle to ensure proper functionality.
