+++ .markdownlint-cli2.jsonc
@@ -27,7 +27,6 @@
     "!playwright-report/**",
     "!playwright-report-dev/**",
     "!playwright-report-preview/**",
-    "!tests/quality/spec_audits/**",
     "!docs/operations/operator-upgrade-notes.md",
     "!docs/agents/issue-tracker.md"
   ]
