+++ tests/sql-integration/requirements-specifications.sqlserver.test.ts
@@ -8,7 +8,7 @@ import {
   createPublishedRequirement,
   createSpecificationFixture,
   useSqlIntegrationDatabase,
-} from './sql-test-database'
+} from './helpers/sql-test-database'
 
 describe('requirements specification mutations', () => {
   const appDb = useSqlIntegrationDatabase()
